feat: Controller에 valid 추가

+++ src/main/java/gift/Controller/ProductAdminController.java
@@ -2,9 +2,11 @@ package gift.Controller;
 
 import gift.Model.ProductModel;
 import gift.Service.ProductService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
@@ -29,20 +31,29 @@ public class ProductAdminController{
         return "product-new";
     }
 
-    @PostMapping("")
-    public String createProduct(@ModelAttribute ProductModel product) {
+    @PostMapping
+    public String createProduct(@Valid @ModelAttribute ProductModel product, BindingResult bindingResult) {
+        if(bindingResult.hasErrors()){
+            return "product-new";
+        }
         productService.saveProduct(product);
         return "redirect:/admin/products";
     }
 
     @GetMapping("/edit/{id}")
     public String showEditForm(@PathVariable("id") Long id, Model model) {
+        if (productService.getProductById(id) == null) {
+            throw new RuntimeException("해당 id의 상품이 존재하지 않습니다");
+        }
         model.addAttribute("product", productService.getProductById(id));
         return "product-edit";
     }
 
     @PutMapping ("/{id}")
-    public String updateProduct(@PathVariable("id") Long id, @ModelAttribute ProductModel product) {
+    public String updateProduct(@Valid @PathVariable("id") Long id, @ModelAttribute ProductModel product, BindingResult bindingResult) {
+        if(bindingResult.hasErrors()){
+            return "product-form";
+        }
         product.setId(id);
         productService.updateProduct(product);
         return "redirect:/admin/products";
@@ -50,6 +61,9 @@ public class ProductAdminController{
 
     @GetMapping("/delete/{id}")
     public String deleteProduct(@PathVariable("id") Long id) {
+        if (productService.getProductById(id) == null) {
+            throw new RuntimeException("해당 id의 상품이 존재하지 않습니다");
+        }
         productService.deleteProduct(id);
         return "redirect:/admin/products";
     }
